Improve AI turn handling and organize card data
