--- src/game/cards.js
@@ -1,810 +1,0 @@
-export const COLORS = {
-  red: {
-    name: 'Fire',
-    theme: 'Aggressive damage and burn spells',
-  },
-  blue: {
-    name: 'Water',
-    theme: 'Control, bounce, and sturdy creatures',
-  },
-  green: {
-    name: 'Grass',
-    theme: 'Growing creatures and life gain',
-  },
-};
-
-let nextInstanceId = 1;
-
-function deepCopyCard(card) {
-  return JSON.parse(JSON.stringify(card));
-}
-
-export function createCardInstance(card) {
-  const copy = deepCopyCard(card);
-  return {
-    ...copy,
-    instanceId: `${card.id}_${nextInstanceId++}`,
-    summoningSickness: card.type === 'creature' && !card.abilities?.haste,
-    damageMarked: 0,
-    activatedThisTurn: false,
-    buffs: [],
-    baseAttack: copy.attack ?? 0,
-    baseToughness: copy.toughness ?? 0,
-  };
-}
-
-export function shuffle(array) {
-  for (let i = array.length - 1; i > 0; i -= 1) {
-    const j = Math.floor(Math.random() * (i + 1));
-    [array[i], array[j]] = [array[j], array[i]];
-  }
-  return array;
-}
-
-function repeatCards(cards) {
-  const duplicated = [];
-  cards.forEach((card) => {
-    duplicated.push(createCardInstance(card));
-    duplicated.push(createCardInstance(card));
-  });
-  return duplicated;
-}
-
-export function buildDeck(color) {
-  const baseCards = CARD_LIBRARY[color];
-  const fullDeck = repeatCards(baseCards);
-  shuffle(fullDeck);
-  return fullDeck;
-}
-
-export const CARD_LIBRARY = {
-  red: [
-    {
-      id: 'red_creature_1',
-      name: 'Flame Sprite',
-      type: 'creature',
-      cost: 1,
-      color: 'red',
-      attack: 2,
-      toughness: 1,
-      abilities: { haste: true },
-      text: '2/1 with haste. Built to deal damage immediately.',
-    },
-    {
-      id: 'red_creature_2',
-      name: 'Ember Runner',
-      type: 'creature',
-      cost: 2,
-      color: 'red',
-      attack: 3,
-      toughness: 1,
-      passive: {
-        type: 'onEnter',
-        description: 'Deal 1 damage to any target when Ember Runner enters.',
-        effect: { type: 'damage', target: 'any', amount: 1 },
-      },
-      text: 'Deals 1 damage when it enters the battlefield.',
-    },
-    {
-      id: 'red_creature_3',
-      name: 'Cinder Scout',
-      type: 'creature',
-      cost: 2,
-      color: 'red',
-      attack: 2,
-      toughness: 2,
-      activated: {
-        type: 'active',
-        cost: 1,
-        oncePerTurn: false,
-        description: 'Spend 1 mana: Target creature gains +1 attack this turn.',
-        effect: { type: 'temporaryBuff', target: 'friendly-creature', attack: 1, toughness: 0 },
-      },
-      text: 'Can pump an ally for the turn.',
-    },
-    {
-      id: 'red_creature_4',
-      name: 'Molten Bruiser',
-      type: 'creature',
-      cost: 3,
-      color: 'red',
-      attack: 4,
-      toughness: 2,
-      text: 'A hard-hitting elemental bruiser.',
-    },
-    {
-      id: 'red_creature_5',
-      name: 'Ashen Phoenix',
-      type: 'creature',
-      cost: 4,
-      color: 'red',
-      attack: 3,
-      toughness: 3,
-      passive: {
-        type: 'onEnter',
-        description: 'Deal 2 damage to target creature when this enters.',
-        effect: { type: 'damage', target: 'enemy-creature', amount: 2 },
-      },
-      text: 'Scorches an enemy on arrival.',
-    },
-    {
-      id: 'red_creature_6',
-      name: 'Inferno Charger',
-      type: 'creature',
-      cost: 4,
-      color: 'red',
-      attack: 5,
-      toughness: 2,
-      abilities: { haste: true },
-      text: 'Big, hasty, and ready to strike the same turn.',
-    },
-    {
-      id: 'red_creature_7',
-      name: 'Lava Tyrant',
-      type: 'creature',
-      cost: 5,
-      color: 'red',
-      attack: 5,
-      toughness: 4,
-      passive: {
-        type: 'onEnter',
-        description: 'Deal 2 damage to the opponent when it enters.',
-        effect: { type: 'damage', target: 'opponent', amount: 2 },
-      },
-      text: 'Burns the enemy directly when it lands.',
-    },
-    {
-      id: 'red_creature_8',
-      name: 'Blazing Duelist',
-      type: 'creature',
-      cost: 3,
-      color: 'red',
-      attack: 2,
-      toughness: 3,
-      passive: {
-        type: 'onAttack',
-        description: 'When it attacks, deal 1 damage to the defending player.',
-        effect: { type: 'damage', target: 'opponent', amount: 1 },
-      },
-      text: 'Chips in extra damage when attacking.',
-    },
-    {
-      id: 'red_creature_9',
-      name: 'Spark Mentor',
-      type: 'creature',
-      cost: 3,
-      color: 'red',
-      attack: 2,
-      toughness: 2,
-      passive: {
-        type: 'static',
-        description: 'Other friendly creatures get +1 attack.',
-        effect: { type: 'globalBuff', scope: 'other-friendly', attack: 1, toughness: 0 },
-      },
-      text: 'Empowers the rest of your squad.',
-    },
-    {
-      id: 'red_creature_10',
-      name: 'Volcano Titan',
-      type: 'creature',
-      cost: 6,
-      color: 'red',
-      attack: 7,
-      toughness: 5,
-      passive: {
-        type: 'onEnter',
-        description: 'Deal 3 damage divided as you choose (auto assigned evenly).',
-        effect: { type: 'splashDamage', amount: 3 },
-      },
-      text: 'Massive finisher blasting enemies on entry.',
-    },
-    {
-      id: 'red_spell_1',
-      name: 'Firebolt',
-      type: 'spell',
-      color: 'red',
-      cost: 1,
-      text: 'Deal 2 damage to a creature.',
-      effects: [{ type: 'damage', target: 'enemy-creature', amount: 2 }],
-    },
-    {
-      id: 'red_spell_2',
-      name: 'Searing Strike',
-      type: 'spell',
-      color: 'red',
-      cost: 2,
-      text: 'Deal 3 damage to any target.',
-      effects: [{ type: 'damage', target: 'any', amount: 3 }],
-    },
-    {
-      id: 'red_spell_3',
-      name: 'Ignite Fury',
-      type: 'spell',
-      color: 'red',
-      cost: 2,
-      text: 'Give a friendly creature +2 attack until end of turn.',
-      effects: [
-        { type: 'temporaryBuff', target: 'friendly-creature', attack: 2, toughness: 0 },
-      ],
-    },
-    {
-      id: 'red_spell_4',
-      name: 'Flame Shield',
-      type: 'spell',
-      color: 'red',
-      cost: 1,
-      text: 'Give a creature +0/+2 permanently.',
-      effects: [{ type: 'buff', target: 'friendly-creature', attack: 0, toughness: 2 }],
-    },
-    {
-      id: 'red_spell_5',
-      name: 'Blazing Inspiration',
-      type: 'spell',
-      color: 'red',
-      cost: 3,
-      text: 'Draw 2 cards.',
-      effects: [{ type: 'draw', amount: 2 }],
-    },
-    {
-      id: 'red_spell_6',
-      name: 'Inferno Wave',
-      type: 'spell',
-      color: 'red',
-      cost: 5,
-      text: 'Deal 3 damage to all enemy creatures and players.',
-      effects: [{ type: 'damageAllEnemies', amount: 3 }],
-    },
-    {
-      id: 'red_spell_7',
-      name: 'Phoenix Call',
-      type: 'spell',
-      color: 'red',
-      cost: 4,
-      text: 'Create a 3/2 Phoenix token with haste.',
-      effects: [
-        {
-          type: 'createToken',
-          token: {
-            name: 'Phoenix Token',
-            attack: 3,
-            toughness: 2,
-            abilities: { haste: true },
-            text: '3/2 haste token.',
-          },
-        },
-      ],
-    },
-    {
-      id: 'red_spell_8',
-      name: 'Scorch Earth',
-      type: 'spell',
-      color: 'red',
-      cost: 3,
-      text: 'Deal 2 damage to all enemy creatures.',
-      effects: [{ type: 'damageAllCreatures', target: 'enemy', amount: 2 }],
-    },
-    {
-      id: 'red_spell_9',
-      name: 'Blitz Formation',
-      type: 'spell',
-      color: 'red',
-      cost: 2,
-      text: 'Two friendly creatures gain haste this turn.',
-      effects: [
-        { type: 'grantHaste', target: 'two-friendly', duration: 'turn' },
-      ],
-    },
-    {
-      id: 'red_spell_10',
-      name: 'Final Spark',
-      type: 'spell',
-      color: 'red',
-      cost: 6,
-      text: 'Deal 5 damage to any target.',
-      effects: [{ type: 'damage', target: 'any', amount: 5 }],
-    },
-  ],
-  blue: [
-    {
-      id: 'blue_creature_1',
-      name: 'Wave Adept',
-      type: 'creature',
-      cost: 1,
-      color: 'blue',
-      attack: 1,
-      toughness: 2,
-      text: 'Resilient initiate of the tide.',
-    },
-    {
-      id: 'blue_creature_2',
-      name: 'Mist Caller',
-      type: 'creature',
-      cost: 2,
-      color: 'blue',
-      attack: 1,
-      toughness: 3,
-      passive: {
-        type: 'onEnter',
-        description: 'Return an enemy creature to its owner\'s hand.',
-        effect: { type: 'bounce', target: 'enemy-creature' },
-      },
-      text: 'Bounces a foe when entering the field.',
-    },
-    {
-      id: 'blue_creature_3',
-      name: 'Harbor Guardian',
-      type: 'creature',
-      cost: 3,
-      color: 'blue',
-      attack: 2,
-      toughness: 4,
-      text: 'Bulky defender to hold the line.',
-    },
-    {
-      id: 'blue_creature_4',
-      name: 'Ripple Mage',
-      type: 'creature',
-      cost: 3,
-      color: 'blue',
-      attack: 2,
-      toughness: 3,
-      activated: {
-        type: 'active',
-        cost: 2,
-        oncePerTurn: true,
-        description: 'Spend 2 mana: Draw a card.',
-        effect: { type: 'draw', amount: 1 },
-      },
-      text: 'Can channel the tide into new spells.',
-    },
-    {
-      id: 'blue_creature_5',
-      name: 'Tide Sculptor',
-      type: 'creature',
-      cost: 4,
-      color: 'blue',
-      attack: 3,
-      toughness: 5,
-      passive: {
-        type: 'static',
-        description: 'Friendly creatures get +0/+1.',
-        effect: { type: 'globalBuff', scope: 'friendly', attack: 0, toughness: 1 },
-      },
-      text: 'Bolsters allied defenses.',
-    },
-    {
-      id: 'blue_creature_6',
-      name: 'Frostbound Levi',
-      type: 'creature',
-      cost: 4,
-      color: 'blue',
-      attack: 3,
-      toughness: 4,
-      passive: {
-        type: 'onEnter',
-        description: 'Tap an enemy creature. It does not untap next turn (simulated by summoning sickness).',
-        effect: { type: 'freeze', target: 'enemy-creature' },
-      },
-      text: 'Locks down an enemy for a turn.',
-    },
-    {
-      id: 'blue_creature_7',
-      name: 'Azure Sentry',
-      type: 'creature',
-      cost: 5,
-      color: 'blue',
-      attack: 3,
-      toughness: 6,
-      text: 'Huge wall of water.',
-    },
-    {
-      id: 'blue_creature_8',
-      name: 'Stormcaller Djinn',
-      type: 'creature',
-      cost: 5,
-      color: 'blue',
-      attack: 4,
-      toughness: 4,
-      passive: {
-        type: 'onEnter',
-        description: 'Draw two cards.',
-        effect: { type: 'draw', amount: 2 },
-      },
-      text: 'Arrives with knowledge from the sky.',
-    },
-    {
-      id: 'blue_creature_9',
-      name: 'Current Warden',
-      type: 'creature',
-      cost: 2,
-      color: 'blue',
-      attack: 1,
-      toughness: 4,
-      text: 'Reliable defensive creature.',
-    },
-    {
-      id: 'blue_creature_10',
-      name: 'Tidal Colossus',
-      type: 'creature',
-      cost: 6,
-      color: 'blue',
-      attack: 6,
-      toughness: 7,
-      passive: {
-        type: 'onEnter',
-        description: 'Return up to two enemy creatures to their owner\'s hand (auto picks strongest).',
-        effect: { type: 'massBounce', amount: 2 },
-      },
-      text: 'Massive finisher that sweeps the board.',
-    },
-    {
-      id: 'blue_spell_1',
-      name: 'Aether Gust',
-      type: 'spell',
-      color: 'blue',
-      cost: 1,
-      text: 'Return a creature to its owner\'s hand.',
-      effects: [{ type: 'bounce', target: 'creature' }],
-    },
-    {
-      id: 'blue_spell_2',
-      name: 'Tidal Insight',
-      type: 'spell',
-      color: 'blue',
-      cost: 2,
-      text: 'Draw two cards.',
-      effects: [{ type: 'draw', amount: 2 }],
-    },
-    {
-      id: 'blue_spell_3',
-      name: 'Mist Barrier',
-      type: 'spell',
-      color: 'blue',
-      cost: 2,
-      text: 'Give a creature +0/+3 until end of turn.',
-      effects: [
-        { type: 'temporaryBuff', target: 'any-creature', attack: 0, toughness: 3 },
-      ],
-    },
-    {
-      id: 'blue_spell_4',
-      name: 'Counter Surge',
-      type: 'spell',
-      color: 'blue',
-      cost: 3,
-      text: 'Return all attacking enemy creatures to hand.',
-      effects: [{ type: 'bounceAttackers' }],
-    },
-    {
-      id: 'blue_spell_5',
-      name: 'Frozen Moment',
-      type: 'spell',
-      color: 'blue',
-      cost: 3,
-      text: 'Freeze an enemy creature for one turn.',
-      effects: [{ type: 'freeze', target: 'enemy-creature' }],
-    },
-    {
-      id: 'blue_spell_6',
-      name: 'Knowledge Dive',
-      type: 'spell',
-      color: 'blue',
-      cost: 4,
-      text: 'Draw three cards.',
-      effects: [{ type: 'draw', amount: 3 }],
-    },
-    {
-      id: 'blue_spell_7',
-      name: 'Illusionary Reinforcements',
-      type: 'spell',
-      color: 'blue',
-      cost: 4,
-      text: 'Create two 2/2 Illusion tokens.',
-      effects: [
-        {
-          type: 'createMultipleTokens',
-          count: 2,
-          token: {
-            name: 'Illusion Token',
-            attack: 2,
-            toughness: 2,
-            text: '2/2 Illusion.',
-          },
-        },
-      ],
-    },
-    {
-      id: 'blue_spell_8',
-      name: 'Geyser Blast',
-      type: 'spell',
-      color: 'blue',
-      cost: 3,
-      text: 'Deal 2 damage to each attacking creature.',
-      effects: [{ type: 'damageAttackers', amount: 2 }],
-    },
-    {
-      id: 'blue_spell_9',
-      name: 'Deep Healing',
-      type: 'spell',
-      color: 'blue',
-      cost: 2,
-      text: 'Remove all damage from a friendly creature and draw a card.',
-      effects: [
-        { type: 'heal', target: 'friendly-creature' },
-        { type: 'draw', amount: 1 },
-      ],
-    },
-    {
-      id: 'blue_spell_10',
-      name: 'Maelstrom Command',
-      type: 'spell',
-      color: 'blue',
-      cost: 5,
-      text: 'Return up to two enemy permanents to hand and draw a card.',
-      effects: [
-        { type: 'massBounce', amount: 2 },
-        { type: 'draw', amount: 1 },
-      ],
-    },
-  ],
-  green: [
-    {
-      id: 'green_creature_1',
-      name: 'Sapling Initiate',
-      type: 'creature',
-      cost: 1,
-      color: 'green',
-      attack: 1,
-      toughness: 2,
-      text: 'Small and sturdy plant warrior.',
-    },
-    {
-      id: 'green_creature_2',
-      name: 'Growth Acolyte',
-      type: 'creature',
-      cost: 2,
-      color: 'green',
-      attack: 2,
-      toughness: 3,
-      passive: {
-        type: 'onEnter',
-        description: 'Give another friendly creature +1/+1.',
-        effect: { type: 'buff', target: 'friendly-creature', attack: 1, toughness: 1, excludeSelf: true },
-      },
-      text: 'Boosts a friend when arriving.',
-    },
-    {
-      id: 'green_creature_3',
-      name: 'Vinebound Bear',
-      type: 'creature',
-      cost: 3,
-      color: 'green',
-      attack: 3,
-      toughness: 3,
-      text: 'Balanced fighter.',
-    },
-    {
-      id: 'green_creature_4',
-      name: 'Thorn Adept',
-      type: 'creature',
-      cost: 3,
-      color: 'green',
-      attack: 2,
-      toughness: 4,
-      activated: {
-        type: 'active',
-        cost: 1,
-        oncePerTurn: false,
-        description: 'Spend 1 mana: Give a creature +0/+2 permanently.',
-        effect: { type: 'buff', target: 'friendly-creature', attack: 0, toughness: 2 },
-      },
-      text: 'Grows allies over time.',
-    },
-    {
-      id: 'green_creature_5',
-      name: 'Barkhide Protector',
-      type: 'creature',
-      cost: 4,
-      color: 'green',
-      attack: 3,
-      toughness: 5,
-      passive: {
-        type: 'static',
-        description: 'Other friendly creatures get +0/+1.',
-        effect: { type: 'globalBuff', scope: 'other-friendly', attack: 0, toughness: 1 },
-      },
-      text: 'Protects the grove.',
-    },
-    {
-      id: 'green_creature_6',
-      name: 'Wild Charger',
-      type: 'creature',
-      cost: 4,
-      color: 'green',
-      attack: 4,
-      toughness: 3,
-      abilities: { haste: true },
-      text: 'Ferocious beast that charges immediately.',
-    },
-    {
-      id: 'green_creature_7',
-      name: 'Elder Treant',
-      type: 'creature',
-      cost: 5,
-      color: 'green',
-      attack: 4,
-      toughness: 6,
-      passive: {
-        type: 'onEnter',
-        description: 'Gain 3 life when it enters.',
-        effect: { type: 'gainLife', amount: 3 },
-      },
-      text: 'Ancient tree that revitalizes you.',
-    },
-    {
-      id: 'green_creature_8',
-      name: 'Nature\'s Herald',
-      type: 'creature',
-      cost: 4,
-      color: 'green',
-      attack: 2,
-      toughness: 4,
-      passive: {
-        type: 'onAttack',
-        description: 'When it attacks, give another friendly creature +1/+1.',
-        effect: { type: 'buff', target: 'friendly-creature', attack: 1, toughness: 1, excludeSelf: true },
-      },
-      text: 'Inspires allies mid-battle.',
-    },
-    {
-      id: 'green_creature_9',
-      name: 'Blooming Hydra',
-      type: 'creature',
-      cost: 5,
-      color: 'green',
-      attack: 5,
-      toughness: 5,
-      activated: {
-        type: 'active',
-        cost: 3,
-        oncePerTurn: true,
-        description: 'Spend 3 mana: Give this creature +2/+2 permanently.',
-        effect: { type: 'selfBuff', attack: 2, toughness: 2 },
-      },
-      text: 'Grows stronger with mana investment.',
-    },
-    {
-      id: 'green_creature_10',
-      name: 'Primal Behemoth',
-      type: 'creature',
-      cost: 6,
-      color: 'green',
-      attack: 7,
-      toughness: 7,
-      passive: {
-        type: 'onEnter',
-        description: 'Give all friendly creatures +1/+1.',
-        effect: { type: 'globalBuff', scope: 'friendly', attack: 1, toughness: 1 },
-      },
-      text: 'Massive finisher that grows the whole team.',
-    },
-    {
-      id: 'green_spell_1',
-      name: 'Wild Growth',
-      type: 'spell',
-      color: 'green',
-      cost: 1,
-      text: 'Give a creature +1/+1.',
-      effects: [{ type: 'buff', target: 'friendly-creature', attack: 1, toughness: 1 }],
-    },
-    {
-      id: 'green_spell_2',
-      name: 'Verdant Blessing',
-      type: 'spell',
-      color: 'green',
-      cost: 2,
-      text: 'Give up to two friendly creatures +1/+1.',
-      effects: [
-        { type: 'multiBuff', count: 2, attack: 1, toughness: 1 },
-      ],
-    },
-    {
-      id: 'green_spell_3',
-      name: 'Nature\'s Resurgence',
-      type: 'spell',
-      color: 'green',
-      cost: 3,
-      text: 'Return a creature from your graveyard to your hand.',
-      effects: [{ type: 'revive', amount: 1 }],
-    },
-    {
-      id: 'green_spell_4',
-      name: 'Life Pulse',
-      type: 'spell',
-      color: 'green',
-      cost: 2,
-      text: 'Gain 4 life.',
-      effects: [{ type: 'gainLife', amount: 4 }],
-    },
-    {
-      id: 'green_spell_5',
-      name: 'Massive Growth',
-      type: 'spell',
-      color: 'green',
-      cost: 3,
-      text: 'Give a creature +3/+3 until end of turn.',
-      effects: [
-        { type: 'temporaryBuff', target: 'friendly-creature', attack: 3, toughness: 3 },
-      ],
-    },
-    {
-      id: 'green_spell_6',
-      name: 'Root Snare',
-      type: 'spell',
-      color: 'green',
-      cost: 2,
-      text: 'Prevent all combat damage to you this turn.',
-      effects: [{ type: 'preventCombatDamage' }],
-    },
-    {
-      id: 'green_spell_7',
-      name: 'Overgrowth Ritual',
-      type: 'spell',
-      color: 'green',
-      cost: 4,
-      text: 'Draw two cards and gain 2 life.',
-      effects: [
-        { type: 'draw', amount: 2 },
-        { type: 'gainLife', amount: 2 },
-      ],
-    },
-    {
-      id: 'green_spell_8',
-      name: 'Pack Summons',
-      type: 'spell',
-      color: 'green',
-      cost: 4,
-      text: 'Create two 3/3 Beast tokens.',
-      effects: [
-        {
-          type: 'createMultipleTokens',
-          count: 2,
-          token: {
-            name: 'Beast Token',
-            attack: 3,
-            toughness: 3,
-            text: '3/3 Beast.',
-          },
-        },
-      ],
-    },
-    {
-      id: 'green_spell_9',
-      name: 'Thrive Together',
-      type: 'spell',
-      color: 'green',
-      cost: 3,
-      text: 'Give a creature +2/+2 and gain 2 life.',
-      effects: [
-        { type: 'buff', target: 'friendly-creature', attack: 2, toughness: 2 },
-        { type: 'gainLife', amount: 2 },
-      ],
-    },
-    {
-      id: 'green_spell_10',
-      name: 'Ancient Awakening',
-      type: 'spell',
-      color: 'green',
-      cost: 5,
-      text: 'Give all friendly creatures +2/+2.',
-      effects: [{ type: 'teamBuff', attack: 2, toughness: 2 }],
-    },
-  ],
-};
-
-export function cardSummary(card) {
-  if (card.type === 'creature') {
-    return `${card.attack}/${card.toughness} creature`;
-  }
-  return 'Spell';
-}
