+++ src/game/cards/blue-cards/blue-creatures.js
@@ -1,0 +1,134 @@
+export const blueCreatures = [
+  {
+    id: 'blue_creature_1',
+    name: 'Wave Adept',
+    type: 'creature',
+    cost: 1,
+    color: 'blue',
+    attack: 1,
+    toughness: 2,
+    text: 'Resilient initiate of the tide.',
+  },
+  {
+    id: 'blue_creature_2',
+    name: 'Mist Caller',
+    type: 'creature',
+    cost: 2,
+    color: 'blue',
+    attack: 1,
+    toughness: 3,
+    passive: {
+      type: 'onEnter',
+      description: "Return an enemy creature to its owner's hand.",
+      effect: { type: 'bounce', target: 'enemy-creature' },
+    },
+    text: 'Bounces a foe when entering the field.',
+  },
+  {
+    id: 'blue_creature_3',
+    name: 'Harbor Guardian',
+    type: 'creature',
+    cost: 3,
+    color: 'blue',
+    attack: 2,
+    toughness: 4,
+    text: 'Bulky defender to hold the line.',
+  },
+  {
+    id: 'blue_creature_4',
+    name: 'Ripple Mage',
+    type: 'creature',
+    cost: 3,
+    color: 'blue',
+    attack: 2,
+    toughness: 3,
+    activated: {
+      type: 'active',
+      cost: 2,
+      oncePerTurn: true,
+      description: 'Spend 2 mana: Draw a card.',
+      effect: { type: 'draw', amount: 1 },
+    },
+    text: 'Can channel the tide into new spells.',
+  },
+  {
+    id: 'blue_creature_5',
+    name: 'Tide Sculptor',
+    type: 'creature',
+    cost: 4,
+    color: 'blue',
+    attack: 3,
+    toughness: 5,
+    passive: {
+      type: 'static',
+      description: 'Friendly creatures get +0/+1.',
+      effect: { type: 'globalBuff', scope: 'friendly', attack: 0, toughness: 1 },
+    },
+    text: 'Bolsters allied defenses.',
+  },
+  {
+    id: 'blue_creature_6',
+    name: 'Frostbound Levi',
+    type: 'creature',
+    cost: 4,
+    color: 'blue',
+    attack: 3,
+    toughness: 4,
+    passive: {
+      type: 'onEnter',
+      description: 'Tap an enemy creature. It does not untap next turn (simulated by summoning sickness).',
+      effect: { type: 'freeze', target: 'enemy-creature' },
+    },
+    text: 'Locks down an enemy for a turn.',
+  },
+  {
+    id: 'blue_creature_7',
+    name: 'Azure Sentry',
+    type: 'creature',
+    cost: 5,
+    color: 'blue',
+    attack: 3,
+    toughness: 6,
+    text: 'Huge wall of water.',
+  },
+  {
+    id: 'blue_creature_8',
+    name: 'Stormcaller Djinn',
+    type: 'creature',
+    cost: 5,
+    color: 'blue',
+    attack: 4,
+    toughness: 4,
+    passive: {
+      type: 'onEnter',
+      description: 'Draw two cards.',
+      effect: { type: 'draw', amount: 2 },
+    },
+    text: 'Arrives with knowledge from the sky.',
+  },
+  {
+    id: 'blue_creature_9',
+    name: 'Current Warden',
+    type: 'creature',
+    cost: 2,
+    color: 'blue',
+    attack: 1,
+    toughness: 4,
+    text: 'Reliable defensive creature.',
+  },
+  {
+    id: 'blue_creature_10',
+    name: 'Tidal Colossus',
+    type: 'creature',
+    cost: 6,
+    color: 'blue',
+    attack: 6,
+    toughness: 7,
+    passive: {
+      type: 'onEnter',
+      description: "Return up to two enemy creatures to their owner's hand (auto picks strongest).",
+      effect: { type: 'massBounce', amount: 2 },
+    },
+    text: 'Massive finisher that sweeps the board.',
+  },
+];
