+++ src/game/cards/blue-cards/blue-spells.js
@@ -1,0 +1,111 @@
+export const blueSpells = [
+  {
+    id: 'blue_spell_1',
+    name: 'Aether Gust',
+    type: 'spell',
+    color: 'blue',
+    cost: 1,
+    text: "Return a creature to its owner's hand.",
+    effects: [{ type: 'bounce', target: 'creature' }],
+  },
+  {
+    id: 'blue_spell_2',
+    name: 'Tidal Insight',
+    type: 'spell',
+    color: 'blue',
+    cost: 2,
+    text: 'Draw two cards.',
+    effects: [{ type: 'draw', amount: 2 }],
+  },
+  {
+    id: 'blue_spell_3',
+    name: 'Mist Barrier',
+    type: 'spell',
+    color: 'blue',
+    cost: 2,
+    text: 'Give a creature +0/+3 until end of turn.',
+    effects: [
+      { type: 'temporaryBuff', target: 'any-creature', attack: 0, toughness: 3 },
+    ],
+  },
+  {
+    id: 'blue_spell_4',
+    name: 'Counter Surge',
+    type: 'spell',
+    color: 'blue',
+    cost: 3,
+    text: 'Return all attacking enemy creatures to hand.',
+    effects: [{ type: 'bounceAttackers' }],
+  },
+  {
+    id: 'blue_spell_5',
+    name: 'Frozen Moment',
+    type: 'spell',
+    color: 'blue',
+    cost: 3,
+    text: 'Freeze an enemy creature for one turn.',
+    effects: [{ type: 'freeze', target: 'enemy-creature' }],
+  },
+  {
+    id: 'blue_spell_6',
+    name: 'Knowledge Dive',
+    type: 'spell',
+    color: 'blue',
+    cost: 4,
+    text: 'Draw three cards.',
+    effects: [{ type: 'draw', amount: 3 }],
+  },
+  {
+    id: 'blue_spell_7',
+    name: 'Illusionary Reinforcements',
+    type: 'spell',
+    color: 'blue',
+    cost: 4,
+    text: 'Create two 2/2 Illusion tokens.',
+    effects: [
+      {
+        type: 'createMultipleTokens',
+        count: 2,
+        token: {
+          name: 'Illusion Token',
+          attack: 2,
+          toughness: 2,
+          text: '2/2 Illusion.',
+        },
+      },
+    ],
+  },
+  {
+    id: 'blue_spell_8',
+    name: 'Geyser Blast',
+    type: 'spell',
+    color: 'blue',
+    cost: 3,
+    text: 'Deal 2 damage to each attacking creature.',
+    effects: [{ type: 'damageAttackers', amount: 2 }],
+  },
+  {
+    id: 'blue_spell_9',
+    name: 'Deep Healing',
+    type: 'spell',
+    color: 'blue',
+    cost: 2,
+    text: 'Remove all damage from a friendly creature and draw a card.',
+    effects: [
+      { type: 'heal', target: 'friendly-creature' },
+      { type: 'draw', amount: 1 },
+    ],
+  },
+  {
+    id: 'blue_spell_10',
+    name: 'Maelstrom Command',
+    type: 'spell',
+    color: 'blue',
+    cost: 5,
+    text: 'Return up to two enemy permanents to hand and draw a card.',
+    effects: [
+      { type: 'massBounce', amount: 2 },
+      { type: 'draw', amount: 1 },
+    ],
+  },
+];
