+++ src/game/cards/blue-cards/index.js
@@ -1,0 +1,6 @@
+import { blueCreatures } from './blue-creatures.js';
+import { blueSpells } from './blue-spells.js';
+
+export const blueCards = [...blueCreatures, ...blueSpells];
+
+export { blueCreatures, blueSpells };
