+++ src/game/cards/green-cards/green-creatures.js
@@ -1,0 +1,142 @@
+export const greenCreatures = [
+  {
+    id: 'green_creature_1',
+    name: 'Sapling Initiate',
+    type: 'creature',
+    cost: 1,
+    color: 'green',
+    attack: 1,
+    toughness: 2,
+    text: 'Small and sturdy plant warrior.',
+  },
+  {
+    id: 'green_creature_2',
+    name: 'Growth Acolyte',
+    type: 'creature',
+    cost: 2,
+    color: 'green',
+    attack: 2,
+    toughness: 3,
+    passive: {
+      type: 'onEnter',
+      description: 'Give another friendly creature +1/+1.',
+      effect: { type: 'buff', target: 'friendly-creature', attack: 1, toughness: 1, excludeSelf: true },
+    },
+    text: 'Boosts a friend when arriving.',
+  },
+  {
+    id: 'green_creature_3',
+    name: 'Vinebound Bear',
+    type: 'creature',
+    cost: 3,
+    color: 'green',
+    attack: 3,
+    toughness: 3,
+    text: 'Balanced fighter.',
+  },
+  {
+    id: 'green_creature_4',
+    name: 'Thorn Adept',
+    type: 'creature',
+    cost: 3,
+    color: 'green',
+    attack: 2,
+    toughness: 4,
+    activated: {
+      type: 'active',
+      cost: 1,
+      oncePerTurn: false,
+      description: 'Spend 1 mana: Give a creature +0/+2 permanently.',
+      effect: { type: 'buff', target: 'friendly-creature', attack: 0, toughness: 2 },
+    },
+    text: 'Grows allies over time.',
+  },
+  {
+    id: 'green_creature_5',
+    name: 'Barkhide Protector',
+    type: 'creature',
+    cost: 4,
+    color: 'green',
+    attack: 3,
+    toughness: 5,
+    passive: {
+      type: 'static',
+      description: 'Other friendly creatures get +0/+1.',
+      effect: { type: 'globalBuff', scope: 'other-friendly', attack: 0, toughness: 1 },
+    },
+    text: 'Protects the grove.',
+  },
+  {
+    id: 'green_creature_6',
+    name: 'Wild Charger',
+    type: 'creature',
+    cost: 4,
+    color: 'green',
+    attack: 4,
+    toughness: 3,
+    abilities: { haste: true },
+    text: 'Ferocious beast that charges immediately.',
+  },
+  {
+    id: 'green_creature_7',
+    name: 'Elder Treant',
+    type: 'creature',
+    cost: 5,
+    color: 'green',
+    attack: 4,
+    toughness: 6,
+    passive: {
+      type: 'onEnter',
+      description: 'Gain 3 life when it enters.',
+      effect: { type: 'gainLife', amount: 3 },
+    },
+    text: 'Ancient tree that revitalizes you.',
+  },
+  {
+    id: 'green_creature_8',
+    name: "Nature's Herald",
+    type: 'creature',
+    cost: 4,
+    color: 'green',
+    attack: 2,
+    toughness: 4,
+    passive: {
+      type: 'onAttack',
+      description: 'When it attacks, give another friendly creature +1/+1.',
+      effect: { type: 'buff', target: 'friendly-creature', attack: 1, toughness: 1, excludeSelf: true },
+    },
+    text: 'Inspires allies mid-battle.',
+  },
+  {
+    id: 'green_creature_9',
+    name: 'Blooming Hydra',
+    type: 'creature',
+    cost: 5,
+    color: 'green',
+    attack: 5,
+    toughness: 5,
+    activated: {
+      type: 'active',
+      cost: 3,
+      oncePerTurn: true,
+      description: 'Spend 3 mana: Give this creature +2/+2 permanently.',
+      effect: { type: 'selfBuff', attack: 2, toughness: 2 },
+    },
+    text: 'Grows stronger with mana investment.',
+  },
+  {
+    id: 'green_creature_10',
+    name: 'Primal Behemoth',
+    type: 'creature',
+    cost: 6,
+    color: 'green',
+    attack: 7,
+    toughness: 7,
+    passive: {
+      type: 'onEnter',
+      description: 'Give all friendly creatures +1/+1.',
+      effect: { type: 'globalBuff', scope: 'friendly', attack: 1, toughness: 1 },
+    },
+    text: 'Massive finisher that grows the whole team.',
+  },
+];
