+++ src/game/cards/green-cards/green-spells.js
@@ -1,0 +1,113 @@
+export const greenSpells = [
+  {
+    id: 'green_spell_1',
+    name: 'Wild Growth',
+    type: 'spell',
+    color: 'green',
+    cost: 1,
+    text: 'Give a creature +1/+1.',
+    effects: [{ type: 'buff', target: 'friendly-creature', attack: 1, toughness: 1 }],
+  },
+  {
+    id: 'green_spell_2',
+    name: 'Verdant Blessing',
+    type: 'spell',
+    color: 'green',
+    cost: 2,
+    text: 'Give up to two friendly creatures +1/+1.',
+    effects: [
+      { type: 'multiBuff', count: 2, attack: 1, toughness: 1 },
+    ],
+  },
+  {
+    id: 'green_spell_3',
+    name: "Nature's Resurgence",
+    type: 'spell',
+    color: 'green',
+    cost: 3,
+    text: 'Return a creature from your graveyard to your hand.',
+    effects: [{ type: 'revive', amount: 1 }],
+  },
+  {
+    id: 'green_spell_4',
+    name: 'Life Pulse',
+    type: 'spell',
+    color: 'green',
+    cost: 2,
+    text: 'Gain 4 life.',
+    effects: [{ type: 'gainLife', amount: 4 }],
+  },
+  {
+    id: 'green_spell_5',
+    name: 'Massive Growth',
+    type: 'spell',
+    color: 'green',
+    cost: 3,
+    text: 'Give a creature +3/+3 until end of turn.',
+    effects: [
+      { type: 'temporaryBuff', target: 'friendly-creature', attack: 3, toughness: 3 },
+    ],
+  },
+  {
+    id: 'green_spell_6',
+    name: 'Root Snare',
+    type: 'spell',
+    color: 'green',
+    cost: 2,
+    text: 'Prevent all combat damage to you this turn.',
+    effects: [{ type: 'preventCombatDamage' }],
+  },
+  {
+    id: 'green_spell_7',
+    name: 'Overgrowth Ritual',
+    type: 'spell',
+    color: 'green',
+    cost: 4,
+    text: 'Draw two cards and gain 2 life.',
+    effects: [
+      { type: 'draw', amount: 2 },
+      { type: 'gainLife', amount: 2 },
+    ],
+  },
+  {
+    id: 'green_spell_8',
+    name: 'Pack Summons',
+    type: 'spell',
+    color: 'green',
+    cost: 4,
+    text: 'Create two 3/3 Beast tokens.',
+    effects: [
+      {
+        type: 'createMultipleTokens',
+        count: 2,
+        token: {
+          name: 'Beast Token',
+          attack: 3,
+          toughness: 3,
+          text: '3/3 Beast.',
+        },
+      },
+    ],
+  },
+  {
+    id: 'green_spell_9',
+    name: 'Thrive Together',
+    type: 'spell',
+    color: 'green',
+    cost: 3,
+    text: 'Give a creature +2/+2 and gain 2 life.',
+    effects: [
+      { type: 'buff', target: 'friendly-creature', attack: 2, toughness: 2 },
+      { type: 'gainLife', amount: 2 },
+    ],
+  },
+  {
+    id: 'green_spell_10',
+    name: 'Ancient Awakening',
+    type: 'spell',
+    color: 'green',
+    cost: 5,
+    text: 'Give all friendly creatures +2/+2.',
+    effects: [{ type: 'teamBuff', attack: 2, toughness: 2 }],
+  },
+];
