+++ src/game/cards/green-cards/index.js
@@ -1,0 +1,6 @@
+import { greenCreatures } from './green-creatures.js';
+import { greenSpells } from './green-spells.js';
+
+export const greenCards = [...greenCreatures, ...greenSpells];
+
+export { greenCreatures, greenSpells };
