+++ src/game/cards/red-cards/index.js
@@ -1,0 +1,6 @@
+import { redCreatures } from './red-creatures.js';
+import { redSpells } from './red-spells.js';
+
+export const redCards = [...redCreatures, ...redSpells];
+
+export { redCreatures, redSpells };
