+++ src/game/cards/red-cards/red-creatures.js
@@ -1,0 +1,141 @@
+export const redCreatures = [
+  {
+    id: 'red_creature_1',
+    name: 'Flame Sprite',
+    type: 'creature',
+    cost: 1,
+    color: 'red',
+    attack: 2,
+    toughness: 1,
+    abilities: { haste: true },
+    text: '2/1 with haste. Built to deal damage immediately.',
+  },
+  {
+    id: 'red_creature_2',
+    name: 'Ember Runner',
+    type: 'creature',
+    cost: 2,
+    color: 'red',
+    attack: 3,
+    toughness: 1,
+    passive: {
+      type: 'onEnter',
+      description: 'Deal 1 damage to any target when Ember Runner enters.',
+      effect: { type: 'damage', target: 'any', amount: 1 },
+    },
+    text: 'Deals 1 damage when it enters the battlefield.',
+  },
+  {
+    id: 'red_creature_3',
+    name: 'Cinder Scout',
+    type: 'creature',
+    cost: 2,
+    color: 'red',
+    attack: 2,
+    toughness: 2,
+    activated: {
+      type: 'active',
+      cost: 1,
+      oncePerTurn: false,
+      description: 'Spend 1 mana: Target creature gains +1 attack this turn.',
+      effect: { type: 'temporaryBuff', target: 'friendly-creature', attack: 1, toughness: 0 },
+    },
+    text: 'Can pump an ally for the turn.',
+  },
+  {
+    id: 'red_creature_4',
+    name: 'Molten Bruiser',
+    type: 'creature',
+    cost: 3,
+    color: 'red',
+    attack: 4,
+    toughness: 2,
+    text: 'A hard-hitting elemental bruiser.',
+  },
+  {
+    id: 'red_creature_5',
+    name: 'Ashen Phoenix',
+    type: 'creature',
+    cost: 4,
+    color: 'red',
+    attack: 3,
+    toughness: 3,
+    passive: {
+      type: 'onEnter',
+      description: 'Deal 2 damage to target creature when this enters.',
+      effect: { type: 'damage', target: 'enemy-creature', amount: 2 },
+    },
+    text: 'Scorches an enemy on arrival.',
+  },
+  {
+    id: 'red_creature_6',
+    name: 'Inferno Charger',
+    type: 'creature',
+    cost: 4,
+    color: 'red',
+    attack: 5,
+    toughness: 2,
+    abilities: { haste: true },
+    text: 'Big, hasty, and ready to strike the same turn.',
+  },
+  {
+    id: 'red_creature_7',
+    name: 'Lava Tyrant',
+    type: 'creature',
+    cost: 5,
+    color: 'red',
+    attack: 5,
+    toughness: 4,
+    passive: {
+      type: 'onEnter',
+      description: 'Deal 2 damage to the opponent when it enters.',
+      effect: { type: 'damage', target: 'opponent', amount: 2 },
+    },
+    text: 'Burns the enemy directly when it lands.',
+  },
+  {
+    id: 'red_creature_8',
+    name: 'Blazing Duelist',
+    type: 'creature',
+    cost: 3,
+    color: 'red',
+    attack: 2,
+    toughness: 3,
+    passive: {
+      type: 'onAttack',
+      description: 'When it attacks, deal 1 damage to the defending player.',
+      effect: { type: 'damage', target: 'opponent', amount: 1 },
+    },
+    text: 'Chips in extra damage when attacking.',
+  },
+  {
+    id: 'red_creature_9',
+    name: 'Spark Mentor',
+    type: 'creature',
+    cost: 3,
+    color: 'red',
+    attack: 2,
+    toughness: 2,
+    passive: {
+      type: 'static',
+      description: 'Other friendly creatures get +1 attack.',
+      effect: { type: 'globalBuff', scope: 'other-friendly', attack: 1, toughness: 0 },
+    },
+    text: 'Empowers the rest of your squad.',
+  },
+  {
+    id: 'red_creature_10',
+    name: 'Volcano Titan',
+    type: 'creature',
+    cost: 6,
+    color: 'red',
+    attack: 7,
+    toughness: 5,
+    passive: {
+      type: 'onEnter',
+      description: 'Deal 3 damage divided as you choose (auto assigned evenly).',
+      effect: { type: 'splashDamage', amount: 3 },
+    },
+    text: 'Massive finisher blasting enemies on entry.',
+  },
+];
