+++ src/game/cards/red-cards/red-spells.js
@@ -1,0 +1,107 @@
+export const redSpells = [
+  {
+    id: 'red_spell_1',
+    name: 'Firebolt',
+    type: 'spell',
+    color: 'red',
+    cost: 1,
+    text: 'Deal 2 damage to a creature.',
+    effects: [{ type: 'damage', target: 'enemy-creature', amount: 2 }],
+  },
+  {
+    id: 'red_spell_2',
+    name: 'Searing Strike',
+    type: 'spell',
+    color: 'red',
+    cost: 2,
+    text: 'Deal 3 damage to any target.',
+    effects: [{ type: 'damage', target: 'any', amount: 3 }],
+  },
+  {
+    id: 'red_spell_3',
+    name: 'Ignite Fury',
+    type: 'spell',
+    color: 'red',
+    cost: 2,
+    text: 'Give a friendly creature +2 attack until end of turn.',
+    effects: [
+      { type: 'temporaryBuff', target: 'friendly-creature', attack: 2, toughness: 0 },
+    ],
+  },
+  {
+    id: 'red_spell_4',
+    name: 'Flame Shield',
+    type: 'spell',
+    color: 'red',
+    cost: 1,
+    text: 'Give a creature +0/+2 permanently.',
+    effects: [{ type: 'buff', target: 'friendly-creature', attack: 0, toughness: 2 }],
+  },
+  {
+    id: 'red_spell_5',
+    name: 'Blazing Inspiration',
+    type: 'spell',
+    color: 'red',
+    cost: 3,
+    text: 'Draw 2 cards.',
+    effects: [{ type: 'draw', amount: 2 }],
+  },
+  {
+    id: 'red_spell_6',
+    name: 'Inferno Wave',
+    type: 'spell',
+    color: 'red',
+    cost: 5,
+    text: 'Deal 3 damage to all enemy creatures and players.',
+    effects: [{ type: 'damageAllEnemies', amount: 3 }],
+  },
+  {
+    id: 'red_spell_7',
+    name: 'Phoenix Call',
+    type: 'spell',
+    color: 'red',
+    cost: 4,
+    text: 'Create a 3/2 Phoenix token with haste.',
+    effects: [
+      {
+        type: 'createToken',
+        token: {
+          name: 'Phoenix Token',
+          attack: 3,
+          toughness: 2,
+          abilities: { haste: true },
+          text: '3/2 haste token.',
+        },
+      },
+    ],
+  },
+  {
+    id: 'red_spell_8',
+    name: 'Scorch Earth',
+    type: 'spell',
+    color: 'red',
+    cost: 3,
+    text: 'Deal 2 damage to all enemy creatures.',
+    effects: [{ type: 'damageAllCreatures', target: 'enemy', amount: 2 }],
+  },
+  {
+    id: 'red_spell_9',
+    name: 'Blitz Formation',
+    type: 'spell',
+    color: 'red',
+    cost: 2,
+    text: 'Two friendly creatures gain haste this turn.',
+    effects: [
+      { type: 'grantHaste', target: 'two-friendly', duration: 'turn' },
+    ],
+  },
+  {
+    id: 'red_spell_10',
+    name: 'Final Spark',
+    type: 'spell',
+    color: 'red',
+    cost: 6,
+    text: 'Deal 5 damage to any target.',
+    effects: [{ type: 'damage', target: 'any', amount: 5 }],
+  },
+];
